backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.27.2
groq==0.37.1
langchain-groq
langchain-core
langgraph
